backend/api/requirements.txt
asgiref==3.4.1
backcall==0.2.0
blessings==1.7
boto3==1.14.20
botocore==1.17.63
click==8.0.3
coverage==5.2
decorator==5.1.0
dj-database-url==0.5.0
Django==3.0.8
django-configurations==2.2
django-cors-headers==3.4.0
django-dry-rest-permissions==1.2.0
django-filter==2.3.0
django-model-utils==4.0.0
django-nose==1.4.6
django-phonenumber-field==5.2.0
django-storages==1.9.1
django-templated-mail==1.1.1
django-unique-upload==0.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
djoser==2.0.3
docutils==0.15.2
drf-flex-fields==0.9.5
factory-boy==3.2.0
Faker==8.14.1
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
ipdb==0.13.3
ipython==7.16.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.2.2
MarkupSafe==2.0.1
mccabe==0.6.1
mkdocs==1.1.2
mock==4.0.2
newrelic==5.14.1.144
nltk==3.6.5
nose==1.3.7
nose-progressive==1.5.2
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.36
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2-binary==2.8.5
ptyprocess==0.7.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.10.0
PyJWT==2.2.0
python-dateutil==2.8.2
pytz==2020.1
PyYAML==6.0
regex==2021.10.8
s3transfer==0.3.7
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
urllib3==1.25.11
wcwidth==0.2.5
